+++ mujoco_warp/_src/types.py
@@ -1725,41 +1725,3 @@ class Data:
   collision_pairid: array("naconmax", wp.vec2i)
   collision_worldid: array("naconmax", int)
   ncollision: array(1, int)
-
-
-@dataclasses.dataclass
-class InverseContext:
-  """Workspace arrays for inverse dynamics."""
-
-  Jaref: wp.array2d(dtype=float)
-  search_dot: wp.array(dtype=float)
-  gauss: wp.array(dtype=float)
-  cost: wp.array(dtype=float)
-  prev_cost: wp.array(dtype=float)
-  done: wp.array(dtype=bool)
-
-
-@dataclasses.dataclass
-class SolverContext:
-  """Workspace arrays for constraint solver."""
-
-  Jaref: wp.array2d(dtype=float)
-  search_dot: wp.array(dtype=float)
-  gauss: wp.array(dtype=float)
-  cost: wp.array(dtype=float)
-  prev_cost: wp.array(dtype=float)
-  done: wp.array(dtype=bool)
-  grad: wp.array2d(dtype=float)
-  grad_dot: wp.array(dtype=float)
-  Mgrad: wp.array2d(dtype=float)
-  search: wp.array2d(dtype=float)
-  mv: wp.array2d(dtype=float)
-  jv: wp.array2d(dtype=float)
-  quad: wp.array2d(dtype=wp.vec3)
-  quad_gauss: wp.array(dtype=wp.vec3)
-  alpha: wp.array(dtype=float)
-  prev_grad: wp.array2d(dtype=float)
-  prev_Mgrad: wp.array2d(dtype=float)
-  beta: wp.array(dtype=float)
-  h: wp.array3d(dtype=float)
-  hfactor: wp.array3d(dtype=float)
